- Giving interserver.lua the boot, merged those properties into connection_properties.lua  - Giving LoginConfig the boot and merging those properties (plus the new ones) into InterServerConfig  - Adding some simple algorithms  - Hitting data providers and some other core classes pretty hard with API redesign - now it shouldn't be possible to mutate the data or config classes from the data providers (where before nothing stopped you) and lots of things are cleaned up and improved as a result  - Changing log types, server types, and a couple other enums to be enum classes  - Removing FindFloorResult and making it a generalized SearchResult  - Quite a bit of style/consistency  - Changing a couple log constants in the config so I could redesign the log loading system (in particular, log_worlds and log_channels are now log_world and log_channel)  - Evaluating many bool functions to see if there's more context that can be given by using enum classes instead of bool (for example, Quest reward iteration was using false to describe "continue iterating" and true to describe "stop iterating" - how you were supposed to know that when just reading the code, I have no idea, but now it has an enum class that gives you StopIterating and ContinueIterating so you can easily tell what the code will do), in general more of this type of evaluation/renaming/etc. needs to be done  - Adding Lua functions that take care of table creation from vectors of int32_t and int8_t  - Log functions were updated so that they will now accept a function that takes an out_stream_t by ref so you don't have to construct it yourself (same with chathandler functions for that matter)
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/branches/buis_vana_overhaul@3288 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/Common/AuthenticationPacket.cpp
@@ -27,6 +27,6 @@ auto AuthenticationPacket::sendPassword(AbstractServerConnection *connection, co
 	packet.add<header_t>(IMSG_PASSWORD);
 	packet.addString(pass);
 	packet.addClassVector<ExternalIp>(extIp);
-	packet.add<int8_t>(connection->getType());
+	packet.add<int8_t>(static_cast<int8_t>(connection->getType()));
 	connection->getSession()->send(packet);
 }
